test_runner: streamline global setup and teardown functions
