+++ lib/internal/test_runner/harness.js
@@ -13,7 +13,7 @@ const {
   createHook,
   executionAsyncId,
 } = require('async_hooks');
-const { relative, resolve, sep } = require('path');
+const { relative } = require('path');
 const {
   codes: {
     ERR_TEST_FAILURE,
@@ -25,45 +25,20 @@ const {
   parseCommandLine,
   reporterScope,
   shouldColorizeTestFiles,
+  setupGlobalSetupTeardownFunctions,
 } = require('internal/test_runner/utils');
 const { queueMicrotask } = require('internal/process/task_queues');
 const { TIMEOUT_MAX } = require('internal/timers');
 const { clearInterval, setInterval } = require('timers');
-const { validatePath } = require('internal/fs/utils');
-const { pathToFileURL } = require('internal/url');
 const { bigint: hrtime } = process.hrtime;
 const resolvedPromise = PromiseResolve();
 const testResources = new SafeMap();
-const esmLoader = require('internal/modules/esm/loader');
-const { kEmptyObject } = require('internal/util');
-const { validateFunction } = require('internal/validators');
-const { getOptionValue } = require('internal/options');
 let globalRoot;
 let globalSetupFunction;
 let globalTeardownFunction;
 let globalSetupExecuted = false;
 
 testResources.set(reporterScope.asyncId(), reporterScope);
-
-async function setupGlobalSetupTeardownFunctions(globalSetupPath, cwd) {
-  let globalSetupFunction;
-  let globalTeardownFunction;
-  if (globalSetupPath) {
-    validatePath(globalSetupPath, 'options.globalSetupPath');
-    const fileURL = pathToFileURL(resolve(cwd, globalSetupPath));
-    const globalSetupModule = await esmLoader.getOrInitializeCascadedLoader()
-        .import(fileURL, pathToFileURL(cwd + sep).href, kEmptyObject);
-    if (globalSetupModule.globalSetup) {
-      validateFunction(globalSetupModule.globalSetup, 'globalSetupModule.globalSetup');
-      globalSetupFunction = globalSetupModule.globalSetup;
-    }
-    if (globalSetupModule.globalTeardown) {
-      validateFunction(globalSetupModule.globalTeardown, 'globalSetupModule.globalTeardown');
-      globalTeardownFunction = globalSetupModule.globalTeardown;
-    }
-  }
-  return { __proto__: null, globalSetupFunction, globalTeardownFunction };
-}
 
 function createTestTree(rootTestOptions, globalOptions) {
   const buildPhaseDeferred = PromiseWithResolvers();
@@ -108,9 +83,7 @@ function createTestTree(rootTestOptions, globalOptions) {
         globalSetupDeferred.resolve();
         return;
       }
-
-      const globalSetupFunctions =
-      await setupGlobalSetupTeardownFunctions(
+      const globalSetupFunctions = await setupGlobalSetupTeardownFunctions(
         globalOptions.globalSetupPath,
         globalOptions.cwd,
       );
@@ -139,9 +112,7 @@ function createTestTree(rootTestOptions, globalOptions) {
   };
 
   harness.resetCounters();
-  if (!getOptionValue('--test')) {
-    harness.waitForGlobalSetup();
-  }
+  harness.waitForGlobalSetup();
   globalRoot = new Test({
     __proto__: null,
     ...rootTestOptions,
@@ -355,11 +326,6 @@ function lazyBootstrapRoot() {
 async function startSubtestAfterBootstrap(subtest) {
   if (subtest.root.harness.buildPromise) {
     if (subtest.root.harness.globalSetupPromise) {
-      if (getOptionValue('--test')) {
-        queueMicrotask(() => {
-          subtest.root.harness.waitForGlobalSetup();
-        });
-      }
       await subtest.root.harness.globalSetupPromise;
       subtest.root.harness.globalSetupPromise = null;
     }
