+++ lib/internal/test_runner/runner.js
@@ -809,9 +809,8 @@ function run(options = kEmptyObject) {
   }
 
   const runChain = async () => {
-    if (root.harness?.waitForGlobalSetup) {
-      await root.harness.waitForGlobalSetup({ __proto__: null, context: root.harness.globalSetupContext });
-      root.harness.globalSetupFunction = null;
+    if (root.harness?.globalSetupPromise) {
+      await root.harness.globalSetupPromise;
     }
     if (typeof setup === 'function') {
       await setup(root.reporter);
@@ -819,10 +818,6 @@ function run(options = kEmptyObject) {
 
     await runFiles();
     postRun?.();
-    if (root.harness?.waitForGlobalTeardown) {
-      await root.harness.waitForGlobalTeardown({ __proto__: null, context: root.harness.globalSetupContext });
-      root.harness.globalTeardownFunction = null;
-    }
     teardown?.();
   };
