+++ lib/internal/test_runner/utils.js
@@ -27,7 +27,7 @@ const {
 } = primordials;
 
 const { AsyncResource } = require('async_hooks');
-const { relative, sep } = require('path');
+const { relative, sep, resolve } = require('path');
 const { createWriteStream } = require('fs');
 const { pathToFileURL } = require('internal/url');
 const { getOptionValue } = require('internal/options');
@@ -41,7 +41,12 @@ const {
   kIsNodeError,
 } = require('internal/errors');
 const { compose } = require('stream');
-const { validateInteger } = require('internal/validators');
+const {
+  validateInteger,
+  validateFunction,
+} = require('internal/validators');
+const { validatePath } = require('internal/fs/utils');
+const { kEmptyObject } = require('internal/util');
 
 const coverageColors = {
   __proto__: null,
@@ -600,6 +605,27 @@ function getCoverageReport(pad, summary, symbol, color, table) {
   return report;
 }
 
+async function setupGlobalSetupTeardownFunctions(globalSetupPath, cwd) {
+  let globalSetupFunction;
+  let globalTeardownFunction;
+  if (globalSetupPath) {
+    validatePath(globalSetupPath, 'options.globalSetupPath');
+    const fileURL = pathToFileURL(resolve(cwd, globalSetupPath));
+    const cascadedLoader = require('internal/modules/esm/loader').getOrInitializeCascadedLoader();
+    const globalSetupModule = await cascadedLoader
+      .import(fileURL, pathToFileURL(cwd + sep).href, kEmptyObject);
+    if (globalSetupModule.globalSetup) {
+      validateFunction(globalSetupModule.globalSetup, 'globalSetupModule.globalSetup');
+      globalSetupFunction = globalSetupModule.globalSetup;
+    }
+    if (globalSetupModule.globalTeardown) {
+      validateFunction(globalSetupModule.globalTeardown, 'globalSetupModule.globalTeardown');
+      globalTeardownFunction = globalSetupModule.globalTeardown;
+    }
+  }
+  return { __proto__: null, globalSetupFunction, globalTeardownFunction };
+}
+
 module.exports = {
   convertStringToRegExp,
   countCompletedTest,
@@ -610,4 +636,5 @@ module.exports = {
   reporterScope,
   shouldColorizeTestFiles,
   getCoverageReport,
+  setupGlobalSetupTeardownFunctions,
 };
