+++ test/parallel/test-runner-run.mjs
@@ -206,7 +206,7 @@ describe('require(\'node:test\').run', { concurrency: true }, () => {
       if (data.name === 'this is a test') {
         t.assert.strictEqual(data.type, 'test');
       }
-    }, 3)); // TODO(pmarchini): this should be 2 but after `await root.harness.waitForGlobalSetup()` it's 3
+    }, 3));
     stream.on('test:dequeue', common.mustCall((data) => {
       if (data.name === 'this is a suite') {
         t.assert.strictEqual(data.type, 'suite');
@@ -214,7 +214,7 @@ describe('require(\'node:test\').run', { concurrency: true }, () => {
       if (data.name === 'this is a test') {
         t.assert.strictEqual(data.type, 'test');
       }
-    }, 3)); // TODO(pmarchini): this should be 2 but after `await root.harness.waitForGlobalSetup()` it's 3
+    }, 3));
 
     // eslint-disable-next-line no-unused-vars
     for await (const _ of stream);
